% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compound_reaction_map.R
\docType{data}
\name{mc_compound_reaction_map}
\alias{mc_compound_reaction_map}
\title{Mapping from compounds to reactions}
\format{A named list of length 12875, where the names correspond to 
\code{mc_compounds$COMPOUND} and the values are a list of compound names 
corresponding to \code{mc_reaction$REACTION}.}
\source{
\url{http://metacyc.ai.sri.com/}
}
\usage{
mc_compound_reaction_map
}
\description{
A list that contains a vector of reaction names for each compound.
}
\seealso{
\code{\link{mc_reactions}}, \code{\link{mc_compounds}}
}
\keyword{datasets}
