% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compounds.R
\docType{data}
\name{mc_compounds}
\alias{mc_compounds}
\title{Compound information from MetaCyc}
\format{A data frame with 13689 rows and 8 variables:
\describe{
  \item{COMPOUND}{ }
  \item{URL}{ }
  \item{COMMON-NAME}{ }
  \item{MONOISOTOPIC-MW}{ }
  \item{MOLECULAR-WEIGHT}{ }
  \item{MF}{ }
  \item{CHEMICAL-FORMULA}{ }
  \item{DBLINKS}{ }
}}
\source{
\url{http://metacyc.ai.sri.com/}
}
\usage{
mc_compounds
}
\description{
A data.frame containing compound information parsed from the MetaCyc database.
}
\keyword{datasets}
