% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_module_reaction_map.R
\docType{data}
\name{mc_module_reaction_map}
\alias{mc_module_reaction_map}
\title{Mapping from modules to reactions}
\format{A named list of length 2572, where the names correspond to 
\code{mc_modules$MODULE} and the values are a list of reaction IDs 
corresponding to \code{mc_reactions$REACTION}.}
\source{
\url{http://metacyc.ai.sri.com/}
}
\usage{
mc_module_reaction_map
}
\description{
A list that contains a vector of reaction IDs for each module.
}
\seealso{
\code{\link{mc_reactions}}, \code{\link{mc_modules}}
}
\keyword{datasets}
