% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_modules.R
\docType{data}
\name{mc_modules}
\alias{mc_modules}
\title{Module information from MetaCyc}
\format{A data frame with 2572 rows and 14 variables:
\describe{
  \item{MODULE}{ }
  \item{URL}{ }
  \item{COMMON-NAME}{ }
  \item{TYPES}{ }
  \item{SUPER-PATHWAYS}{ }
  \item{SUB-PATHWAYS}{ }
  \item{REACTION-LIST}{ }
  \item{SPECIES}{ }
  \item{TAXONOMIC-RANGE}{ }
  \item{PREDECESSORS}{ }
  \item{PATHWAY-LINKS}{ }
  \item{COMMENT}{ }
  \item{DBLINKS}{ }
  \item{IS_SUPER_PATHWAY}{ }
}}
\source{
\url{http://metacyc.ai.sri.com/}
}
\usage{
mc_modules
}
\description{
A data.frame containing module information parsed from the MetaCyc database.
Note that in MetaCyc, these are called Pathways, however, "Modules" only include
ones that are not super-pathways.
}
\keyword{datasets}
