% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_reaction_compound_map.R
\docType{data}
\name{mc_reaction_compound_map}
\alias{mc_reaction_compound_map}
\title{Mapping from reactions to compounds}
\format{A named list of length 15311, where the names correspond to 
\code{mc_reaction$REACTION} and the values are a list of compound names 
corresponding to \code{mc_compounds$COMPOUND}.}
\source{
\url{http://metacyc.ai.sri.com/}
}
\usage{
mc_reaction_compound_map
}
\description{
A list that contains a vector of compound names for each reaction.
}
\seealso{
\code{\link{mc_reactions}}, \code{\link{mc_compounds}}
}
\keyword{datasets}
