% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_reactions.R
\docType{data}
\name{mc_reactions}
\alias{mc_reactions}
\title{Reaction information from MetaCyc}
\format{A data frame with 15311 rows and 11 variables:
\describe{
  \item{REACTION}{ }
  \item{URL}{ }
  \item{SYSTEMATIC-NAME}{ }
  \item{EC-NUMBER}{ }
  \item{IN-PATHWAY}{ }
  \item{REACTION-LIST}{ }
  \item{LEFT}{ }
  \item{RIGHT}{ }
  \item{REACTION-DIRECTION}{ }
  \item{COMPOUNDS}{ }
  \item{DBLINKS}{ }
}}
\source{
\url{http://metacyc.ai.sri.com/}
}
\usage{
mc_reactions
}
\description{
A data.frame containing reaction information parsed from the MetaCyc database.
}
\keyword{datasets}
